bugfix: cstart64: update selectors after switching GDT on BSP
There is no guarantee on the exact values in the selectors of data segments,
including DS, ES, FS, GS and SS, at the time an OS is invoked, according to
Section 3.2, Multiboot Specification 0.6.96. Not re-initializing the selectors
can lead to segmentation faults the first time a reload of segment base address
is needed.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>