apic: switch to x2apic mode before accessing any lapic register
The lapic may not be in xapic mode when the unit test starts. To get neutral on
the apic mode on startup, this patch enables x2apic if LAPIC is yet in xapic
mode, before any lapic register is accessed, and stick to the x2apic way of
accessing the registers afterwards.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>